pyproject.toml
[project]
name = "chat-with-video"
version = "0.1.0"
description = "MiniCPM-V 视频聊天 Demo - Intel GPU版本"
readme = "README.md"
requires-python = ">=3.11"
authors = [
    {name = "PCcoding666", email = "e1143754@u.nus.edu"}
]
keywords = ["video", "chat", "AI", "multimodal", "intel-gpu", "minicpm-v"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Multimedia :: Video",
]
dependencies = [
    "decord>=0.6.0",
    "numpy>=1.24.0,<2.0.0",
    "opencv-python>=4.10.0",
    "pillow>=10.0.0",
    "scipy>=1.14.0",
    "transformers>=4.50.0",
    "gradio>=4.44.0",
    "pyyaml>=6.0",
    # Intel XPU PyTorch需要从特定index安装
    "torch",
    "torchvision",
    "torchaudio",
    "accelerate>=1.10.1",
    "bitsandbytes>=0.47.0",
]

[project.urls]
Homepage = "https://github.com/PCcoding666/Chat_with_video"
"Bug Tracker" = "https://github.com/PCcoding666/Chat_with_video/issues"

[project.scripts]
video-chat = "main:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/chat_with_video"]

[tool.uv]
index-url = "https://pypi.org/simple/"
extra-index-url = ["https://download.pytorch.org/whl/xpu"]

[tool.black]
line-length = 88
target-version = ['py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
